refactor: give the data model a crate of its own
