+++ Cargo.toml
@@ -1,6 +1,17 @@
-[package]
-name = "cydonia"
-description = "Desktop workspace for the ACP agents you run, keeping what they produce as files on your disk"
+# The workspace. Two crates: the app, and the data model it is built on.
+#
+# `schema` is the half that has to be readable from outside — the shapes the
+# app-level MCP server answers with, and what anyone building on cydonia
+# depends on. Keeping it a crate of its own is what makes that a contract
+# rather than a promise: nothing in it can reach for gpui, so nothing about
+# the window can leak into a shape a client reads.
+[workspace]
+members = ["crates/cydonia", "crates/schema"]
+resolver = "3"
+
+# One version for both, bumped together. They ship as a pair — a schema the
+# app cannot read is not a release of anything.
+[workspace.package]
 version = "0.1.1"
 edition = "2024"
 # The floor the edition itself sets. Nothing here reaches past it on purpose.
@@ -8,66 +19,18 @@ rust-version = "1.85"
 license = "MIT"
 repository = "https://github.com/crabtalk/cydonia"
 homepage = "https://cydonia.sh"
-readme = "README.md"
-keywords = ["acp", "agents", "desktop", "gpui", "workspace"]
-categories = ["gui", "development-tools"]
-# What a `cargo install` needs and nothing else. The default is every tracked
-# file, which would ship the website and the bundling plumbing to anyone who
-# installs the binary. `assets/` is absent by design — the logo is downloaded,
-# never committed, and `mark()` already holds the case where it is missing.
-include = ["/src", "/bin", "/build.rs", "/README.md", "/LICENSE"]
 
-[[bin]]
-name = "cydonia"
-path = "bin/main.rs"
-
-[dependencies]
-# bezel — the component library the GUI is built on. Published as `bezel-*`;
-# the keys here are what the source imports. `platform` carries gpui_platform,
-# which opens the window: gpui core knows no platform of its own.
-bezel = { version = "0.1.8", features = ["platform"] }
-editor = { version = "0.1.8", package = "bezel-editor" }
-markdown = { version = "0.1.8", package = "bezel-markdown" }
-syntax = { version = "0.1.8", package = "bezel-syntax" }
-
-# cacp — our ACP implementation, taken from the registry like any other
-# dependency now that it is published and settled.
-cacp = { version = "0.1.0", default-features = false, features = ["client"] }
-cacp-agents = "0.1.0"
-
-# crates.io
+# What both crates take, so a version bump lands in one place.
+[workspace.dependencies]
 anyhow = "1"
-dirs = "6"
-# `Shared`, which is the task an in-flight cover is held as. gpui builds its own
-# image cache on it and compiles the crate either way, so this names what is
-# already in the graph rather than adding to it.
-futures = "0.3"
-# Resizes a picked cover on the way in. Already in the graph — gpui decodes
-# every image it paints with it — so this names a crate that is compiled either
-# way rather than adding one.
-image = "0.25"
-# The cover cache's recency list. Zero dependencies of its own — the whole
-# crate is the map and the order, which is exactly what is missing from gpui's
-# own asset cache.
-lru = "0.18"
-# The filesystem watch behind a project's `.cydonia/`. One backend per
-# platform — FSEvents here, inotify on Linux — behind one interface, which is
-# the whole reason to take a dependency for this rather than poll.
-notify = "8"
-# The project database. `bundled` compiles SQLite in: the store is a file we
-# hand to the user, so it cannot depend on whichever sqlite3 a machine ships.
-rusqlite = { version = "0.40", features = ["bundled"] }
+# cacp — our ACP implementation. `schema` takes it for the one tool-call enum a
+# transcript carries; the app takes the client with it.
+cacp = { version = "0.1.0", default-features = false }
 serde = { version = "1", features = ["derive"] }
 # An archived transcript is an array of tagged enum variants, which is the one
 # shape `toml` cannot round-trip.
 serde_json = "1"
-# cacp spawns its read and write loops with `tokio::spawn`, so the connection
-# needs a runtime of its own — gpui's executor is smol's.
-tokio = { version = "1", features = ["process", "rt-multi-thread", "sync", "time"] }
 toml = "0.8"
-# Installing an agent edits settings.toml in place. `toml` would round-trip
-# it through a value tree and drop every comment the file was written with.
+# Editing a settings or properties file in place. `toml` would round-trip it
+# through a value tree and drop every comment the file was written with.
 toml_edit = "0.22"
-# Downloads the registry's agent icons. The same client cacp-agents fetches
-# the catalog with, so this adds no second HTTP stack.
-ureq = "3"
