+++ crates/cydonia/src/assets.rs
