+++ crates/cydonia/src/model/cover.rs
