+++ crates/cydonia/src/model/mod.rs
@@ -3,11 +3,8 @@
 //! from.
 
 pub mod article;
-pub mod board;
 pub mod cover;
 pub mod project;
-pub mod properties;
-pub mod record;
 pub mod session;
 pub mod settings;
 pub mod state;
