+++ crates/cydonia/src/view/sidebar.rs
